Minimal task met! Renderer 'Complete'
The engine is now 'MVT Complete' for the renderer (Meaning Minimal
Viable Task) and so I'm moving on to other basic foundations I need.
The next task to tackle is events management as this allows for user
input which in turn hooks into camera movement. In addition, having my
own fully designed events manager should make it easier to make put in
networking should I want to.