Changing vector magnitude calc
Small change to change over the use of 'Pow' to a straight
multiplication of each integer with itself, as suggested by Pete.
Looking at the disassembly it looks like the straight multiplication
takes a whole 3 instructions versus the 6 to perform the function call
and however many behind the scenes to do it.