Changing to derivative Events vs 'Packet' style
Changed over the Event system to begin using a set of derivative classes
over the 'Packet' style that I wanted to use. The latter style using a
void pointer to make it so generic it could be used anywhere. This gave
rise to issues where the cleanup upon deletion was an awful pain to
achieve and the only way I saw to deal with it was rather smelly
(Involving reallocating the memory block for the void* to something else
before deleting it).
Also some extremely early work on a logging system. Which reminds me,
must look into the Google logging system for inspiration.