More Gamelog work/Circular Includes
It's a problem I'm continuing to have and it irritates me. I think they
are called circular includes and if they are not, they should be. Need
to find a way of safeguarding against it as it's a goddamn pain in the
ass to deal with.
Also there's a little more work with the game log that's gone in.

Goddamn I love singletons.