Physics Collision Detection Added!
Added a collision detector. It's pretty basic and I don't believe I'm
close to getting something that handles everything I want it to. Perhaps
I'm asking for too much but, who knows.

In any case, the code will run just fine in this update. The camera has
moved to be side on to watch the off-centre collision of the two boxes
as they accidentally collide several more times and end up firing off
more times than I want. The actual addition of forces to the boxes is
placeholder.