Historical Commit: Physics
This is a historical commit, basically. I'm going to revamp the whole of
the physics system as I dislike it and it's a whole mishmash of details.
I'm having implementing a constant collision detection but it's just
causing me more of a headache than I would like.

If I ever need a 2D physics simulator though, this might be a good
choice when most things will be perfectly inelastic collisions.