More physics work.
So I've done a little more on the physics end, hard to write/architect
code when you're a little burnt out. Currently trying to generalise it
for the possible collisions I want to handle at the moment. Of course I
choose one of the more complex ones to do, which is box v box. Might
switch over to sphere/sphere at some point to recheck collision
resolutions and ensure that's all peachy but I'm unsure at the moment.