LocalToWorld Matrix/Quaternion math
Finished adding the LocalToWorld matrix to the entity object, as well as
finishing/polishing off the quaternion maths behind it.

Also realised that I was performing _CMATHS_ cos/sin functions with
degrees instead of radians. Thus I was getting wrong answers for things.

WolframAlpha is a god when it comes to proofing these functions.