Working Physics model
So this commit adds a, working, physics model with the current
architecture I've set out. It still needs a lot of work to deal with
many other situations and objects but it's a plain start to work from.
This mainly brings with it the Seperating Axis Theory tests to detect
two objects which may be colliding. I also want to check out GJK but,
for the moment, I'm content with this method.
There's also a small fix for the log file not being removed when
starting a fresh instance of the engine. Before it was merely moving the
data to a new location and then appending new data to it, which was
obviously wrong.
In addition, always re-read the math. One of the frustrating issues I
came across was where I had made a rookie mistake in simplifying some of
the maths involved.