Working camera system
We now have a working camera system! Yay!

So to move we use W,A,S and D in the normal FPS movement style. Q and E
rotate the camera along its look axis (Z in our case) anti-clockwise and
clockwise respectively. In addition, our mouse movement will cause the
camera to tilt and pan in the usual manner.

Next up is capturing the mouse so that it can't go outside the borders
(Or any other method that allows me to rotate freely) and possibly
adding a build script that bundles up the build into a Zip file. That
way I can have a ready-to-go version of the current executable in the
repo.